Add OpenCloseIterator and refactor LocalNode to make it reusable
