+++ sql/core/src/main/scala/org/apache/spark/sql/execution/local/FilterNode.scala
@@ -24,24 +24,28 @@ import org.apache.spark.sql.catalyst.expressions.codegen.GeneratePredicate
 
 case class FilterNode(condition: Expression, child: LocalNode) extends UnaryLocalNode {
 
-  private[this] var predicate: (InternalRow) => Boolean = _
-
   override def output: Seq[Attribute] = child.output
 
-  override def open(): Unit = {
-    child.open()
-    predicate = GeneratePredicate.generate(condition, child.output)
-  }
+  override def execute(): OpenCloseRowIterator = new OpenCloseRowIterator {
 
-  override def next(): Boolean = {
-    var found = false
-    while (child.next() && !found) {
-      found = predicate.apply(child.get())
+    private val childIter = child.execute()
+
+    private val predicate = GeneratePredicate.generate(condition, child.output)
+
+    override def open(): Unit = childIter.open()
+
+    override def close(): Unit = childIter.close()
+
+    override def getRow: InternalRow = childIter.getRow
+
+    override def advanceNext(): Boolean = {
+      var found = false
+      while (!found && childIter.advanceNext()) {
+        found = predicate.apply(childIter.getRow)
+      }
+      found
     }
-    found
+
   }
 
-  override def get(): InternalRow = child.get()
-
-  override def close(): Unit = child.close()
 }
